refactor: add some utils functions
