+++ lib/knife/tag_utils.js
@@ -105,6 +105,18 @@ function get_classes(class_attribute) {
     .split(/\s+/);
 }
 
+/**
+ *
+ * @param {*} node
+ * @returns {String} name of the tag
+ */
+function node_tag_name(node) {
+  if (node.type === "text") {
+    return "Text Node"; // get text node content but truncate ?
+  }
+  return node.name;
+}
+
 module.exports = {
   has_non_empty_attribute,
   attribute_value,
@@ -114,5 +126,6 @@ module.exports = {
   get_attribute,
   has_attribute,
   attribute_has_value,
-  get_classes
+  get_classes,
+  node_tag_name
 };
