+++ lib/knife/text_node_utils.js
@@ -1,3 +1,5 @@
+const { is_text_node } = require("./tag_utils");
+
 /**
  *
  *
@@ -31,6 +33,15 @@ function get_lines(node, include_EOL = false) {
   return lines;
 }
 
+function is_newline_only(node) {
+  if (!is_text_node(node)) {
+    return false;
+  }
+  const text = node.data || "";
+  return /^([\n\r]+[\t ]*)+$/.test(text);
+}
+
 module.exports = {
-  get_lines
+  get_lines,
+  is_newline_only
 };
