+++ lib/rules/indent-style/index.js
@@ -1,4 +1,8 @@
-const { is_text_node } = require("../../knife/tag_utils");
+const {
+  is_text_node,
+  node_tag_name
+} = require("../../knife/tag_utils");
+const { is_newline_only } = require("../../knife/text_node_utils");
 const { create_list_value_validator } = require("../../validate_option");
 const { create_number_validator } = require("../../validate_option");
 
@@ -89,11 +93,6 @@ function check_indent_width_close({ open, close }) {
   return open.loc.start.column === close.loc.start.column;
 }
 
-function is_newline_only(node) {
-  const text = node.data || "";
-  return /^([\n\r]+[\t ]*)+$/.test(text);
-}
-
 function check_indent_style(node, indent_style) {
   if (is_parent_open_close_same_line(node)) {
     return true;
@@ -115,6 +114,7 @@ function check_indent_style(node, indent_style) {
       return /(\t+ +| +\t)/.test(text) === false;
   }
 }
+
 function indent_style_used(node, indent_style) {
   let text = node.indent;
   text = text.replace(/^\n*/, "");
@@ -125,13 +125,6 @@ function indent_style_used(node, indent_style) {
     return "spaces";
   }
   return "mixed";
-}
-
-function nodeTagName(node) {
-  if (node.type === "text") {
-    return "Text Node"; // get text node content but truncate ?
-  }
-  return node.name;
 }
 
 function check_node_indent(node, indent, expected_indent_width, report) {
@@ -145,7 +138,7 @@ function check_node_indent(node, indent, expected_indent_width, report) {
         position: node.loc,
         meta: {
           data: {
-            tagName: nodeTagName(node),
+            tagName: node_tag_name(node),
             expected_indentation: expected_indent_width,
             current_indentation: node.loc.start.column - 1,
             isClose: false
@@ -159,7 +152,7 @@ function check_node_indent(node, indent, expected_indent_width, report) {
         position: node.close.loc,
         meta: {
           data: {
-            tagName: nodeTagName(node),
+            tagName: node_tag_name(node),
             expected_indentation: expected_indent_width,
             current_indentation: node.close.loc.start.column - 1,
             isClose: true
@@ -175,7 +168,7 @@ function check_node_indent(node, indent, expected_indent_width, report) {
       position: node.loc,
       meta: {
         data: {
-          tagName: nodeTagName(node),
+          tagName: node_tag_name(node),
           expected_type: indent.style,
           current_type: indent_style_used(node),
           expected_indentation: expected_indent_width,
